+++ src/Worker/AzureManaged/DurableTaskSchedulerWorkerExtensions.cs
@@ -3,6 +3,7 @@
 
 using Azure.Core;
 using Microsoft.DurableTask.Worker.Grpc;
+using Microsoft.DurableTask.Worker.Grpc.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
@@ -89,6 +90,12 @@ public static class DurableTaskSchedulerWorkerExtensions
             .Configure(additionalConfig ?? (_ => { }))
             .ValidateDataAnnotations();
 
+        builder.Services.AddOptions<DurableTaskWorkerOptions>(builder.Name)
+           .Configure(options =>
+           {
+               options.EnableEntitySupport = true;
+           });
+
         builder.Services.TryAddEnumerable(
             ServiceDescriptor.Singleton<IConfigureOptions<GrpcDurableTaskWorkerOptions>, ConfigureGrpcChannel>());
         builder.UseGrpc(_ => { });
@@ -117,6 +124,7 @@ public static class DurableTaskSchedulerWorkerExtensions
         {
             DurableTaskSchedulerWorkerOptions source = schedulerOptions.Get(name ?? Options.DefaultName);
             options.Channel = source.CreateChannel();
+            options.ConfigureForAzureManaged();
         }
     }
 }